:recycle: Refactor: FeedItem 컴포넌트 분리
