+++ src/components/Feed/Feed.module.css
@@ -2,24 +2,15 @@
   @apply sr-only;
 }
 
-.feedImage {
-  @apply bg-slate-200 w-24 h-24;
-}
-.feedInfo {
-  @apply flex justify-between mt-2 w-24;
-}
-.userWrapper {
-  @apply flex gap-1 justify-between;
+.headerSection {
+  @apply flex justify-between px-2 mt-2;
 }
 
-.userImg {
-  @apply bg-gray-300 rounded-full w-6 h-6 self-center;
+.newLabel {
+  @apply font-semibold text-red-600;
 }
 
-.speechBubbleHead {
-  @apply relative bg-white border border-black;
-}
-.speechBubbleBody {
-  @apply absolute inset-0 transform translate-x-1/2 -translate-y-1/2 -rotate-45;
-  @apply bg-white border border-t-black border-r-black border-l-white border-b-white  w-2 h-2;
+.feedWrapper {
+  @apply grid  grid-cols-3 gap-x-2 gap-y-6 md:grid-cols-4;
+  @apply min-w-[320px] w-full  mt-4 mx-auto;
 }
